chatapi-palm/requirements.txt
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudevents==1.9.0
cryptography==41.0.3
deprecation==2.1.0
Flask==2.3.3
functions-framework==3.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-aiplatform==1.28.0
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-logging==3.6.0
google-cloud-resource-manager==1.10.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
ndg-httpsclient==0.5.1
oauth2client==4.1.3
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
urllib3==1.26.16
watchdog==3.0.0
Werkzeug==2.3.7
